engine/requirements.txt
openai==2.8.0
python-dotenv==1.2.1
requests==2.32.4
faiss-cpu==1.13.2
numpy==2.3.1
sentence-transformers==5.2.2
torch==2.10.0
rank-bm25==0.2.2
joblib==1.5.3
gitpython==3.1.46
slowapi==0.1.9
pytest==9.0.2
beautifulsoup4==4.14.3
fastapi==0.116.1
uvicorn==0.41.0
pydantic==2.11.7
pydantic-settings==2.7.0
guardrails-ai==0.9.2
websockets==13.1
mcp>=1.0.0
watchfiles==1.0.5
